move infinite scroll in its own module
