+++ ui/site/src/index.ts
@@ -2,4 +2,5 @@ import './standalones/util';
 import './timeago';
 import './standalones/trans';
 import './socket';
+import './infinite-scroll';
 import './main';
